(Refactored latin as well)
