+++ pie_extended/models/lasla/__init__.py
@@ -1,2 +1,2 @@
 from pie_extended.models.lasla.consts import DOWNLOADS, Models, Disambiguator, addons, DESC
-from pie_extended.models.lasla.classes import get_iterator_and_processor
+from pie_extended.models.lasla.get import get_iterator_and_processor
